+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 - ADDED
 - REMOVED
+  - `{component}-trigger` classes. Now only `data-target` is needed.
 - IMPROVED
   - Tooltip timeouts handling
   - Lightbox rework for better behavior
+  - Collapsible code
 - BUGFIXES
 
 # 2.0.0-beta.5 - 2022-01-04